// Downloads/dpm-ui-web-and-mobile-fixes 1/dpm-ui-web-and-mobile-fixes/web-apps/consumer-portal/src/types/quoteAndBuy.ts
  export interface DriverDetails {
    maritalStatusCd?: number;
    childrenBelow16?: number;
    driverRelationship?: string | null;
    educationLevel?: number;
    licenseCountry?: string;
    trafficViolations?: string;
    healthConditions?: string;
    najmCaseDetails?: any[];
    validDrivingLicenses?: any[];
  }
  
  export interface DriverDetailsResponseData {
    driverID: string;
    driverIDType: string | null;
    driverName: string;
    driverNameArabic: string;
    dateofBirth: string;
    dateofBirthH: string;
    nationality: string;
    gender: string;
    occupation: string;
    workCompanyName: string | null;
    workCityCode: string | null;
    homeCityCode: string | null;
    homeAddress: string;
    licenseType: number | null;
    licenseYear: number | null;
    licenseExpiryDateH: string | null;
    idIssuePlaceCode: string;
    noOfAccidents: number;
    noOfClaims: number;
    unitNo: string | null;
    buildingNumber: string;
    streetName: string;
    district: string;
    city: string;
    additionalNumber: string;
    postalCode: string;
    healthConditions: string | null;
    trafficViolations: string | null;
    najmCaseDetails?: {[key:string]: string}[];
    validDrivingLicenses?: string[];
    mainDriverInd: string;
    ncdFreeYears: string | null;
    ncdReferenceNo: string | null;
    relation?: string | null;
    relationship?: string | null;
    additionalDriverDetails?: DriverDetails
  }

  export interface OwnerDetailsResponseData {
    message: string,
    isValid: boolean,
    referenceNo: number,
    sessionSecretId: string,
    ownerFullNameEnglish: string,
    ownerFullNameArabic: string,
    ownerDobG: string,
    ownerDobH: string,
    gender: string,
    nationality: string,
    nationalityCode: string,
    mobileNumber?: string,
    ownerId?: string
  }


  export interface DriverDetailsMappingFields {
    mainDriverInd: string;
    usagePercentage: number;
    nationality: string;
    driverIDType: number;
    driverID: string;
    driverName: string;
    driverNameArabic: string;
    relation: number;
    dateOfBirth: string;
    dateOfBirthH: string;
    gender: string;
    occupation?: string;
    educationLevel: number;
    maritalStatusCd: number;
    childrenBelow16: number;
    workCompanyName?: string | null;
    workCityCode?: string;
    homeCityCode?: string;
    homeAddress?: string;
    licenseType?: number;
    licenseYear: number;
    licenseExpiryDateH?: string | null;
    idIssuePlaceCode: string;
    ncdReference?: string;
    noOfAccidents: number;
    noOfClaims: number;
    unitNo?: string;
    buildingNumber: string;
    streetName: string;
    district: string;
    city: string;
    additionalNumber: string;
    postalCode: string;
    healthConditions?: string;
    trafficViolations?: string;
    najmCaseDetails?: any[];
    validDrivingLicenses?: any[];
  }